View entries as QR codes
Add a feature which allows an entry's username and password to be viewed as
QR codes. The QR codes are hidden until a button is clicked, which reveals
them for a limited time, the timeout being settable in the settings dialog.

This is useful for being able to quickly enter long and secure passwords
into mobile devices.